Put time filters in common
