+++ cli/src/pcluster/models/imagebuilder_resources.py
@@ -12,8 +12,6 @@
 # This module contains all the classes representing the Resources objects.
 # These objects are obtained from the configuration file through a conversion based on the Schema classes.
 #
-from datetime import datetime
-
 from pcluster.aws.aws_api import AWSApi
 from pcluster.aws.aws_resources import StackInfo
 from pcluster.aws.common import AWSClientError
@@ -142,51 +140,3 @@ class ImageBuilderStack(StackInfo):
             except (AWSClientError, KeyError):
                 return None
         return None
-
-
-class ExportImageLogsFiltersParser:
-    """Class to manage export image logs filters."""
-
-    def __init__(
-        self,
-        log_group_name: str,
-        start_time: str = None,
-        end_time: str = None,
-    ):
-        self._log_group_name = log_group_name
-        try:
-            self._start_time = isoformat_to_epoch(start_time) if start_time else None
-            self.end_time = isoformat_to_epoch(end_time) if end_time else int(datetime.now().timestamp() * 1000)
-        except Exception as e:
-            raise FiltersParserError(f"Unable to parse time. It must be in ISO8601 format. {e}")
-
-    @property
-    def start_time(self):
-        """Get start time filter."""
-        if not self._start_time:
-            try:
-                self._start_time = AWSApi.instance().logs.describe_log_group(self._log_group_name).get("creationTime")
-            except AWSClientError as e:
-                raise FiltersParserError(
-                    f"Unable to retrieve creation time of log group {self._log_group_name}, {str(e)}"
-                )
-        return self._start_time
-
-    @start_time.setter
-    def start_time(self, value):
-        """Set start_time value."""
-        self._start_time = value
-
-    def validate(self):
-        """Check filter consistency."""
-        if self.start_time >= self.end_time:
-            raise FiltersParserError("Start time must be earlier than end time.")
-
-        event_in_window = AWSApi.instance().logs.filter_log_events(
-            log_group_name=self._log_group_name, start_time=self.start_time, end_time=self.end_time
-        )
-        if not event_in_window:
-            raise FiltersParserError(
-                f"No log events in the log group {self._log_group_name} in interval starting "
-                f"at {self.start_time} and ending at {self.end_time} "
-            )
